\name{plot.Kset}
\alias{plot.Kset}
\title{
Plot the results of the \code{Kset} function
}
\description{
Same as \code{plot.Knet}, but for output from the \code{Kset} function.
}
\references{
Cornish, A.J. and Markowetz, F. (2014) \emph{SANTA: Quantifying the Functional Content of Molecular Networks.}. PLOS Computational Biology. 10:9, e1003808.
}
\author{
Alex J. Cornish and Florian Markowetz
}

